feat(object): generate dispersed object key (#726)
* feat(object): generate dispersed object key


* fix(object): fix example


---------

+++ components/object-storage/src/lib.rs
@@ -47,7 +47,7 @@ pub trait ObjectManager {
 }
 
 pub struct Owner {
-    pub epoch: u32,
+    pub epoch: u16,
     pub start_offset: u64,
 }
